lambda-functions/processImage/requirements.txt
boto3==1.42.24
boto3-stubs[s3]==1.42.24
boto3-stubs[sqs]==1.42.24
Pillow==12.1.0